+++ crates/consensus/src/receipt/receipts.rs
@@ -1,13 +1,12 @@
-use crate::{
-    receipt::{Eip658Value, RlpDecodableReceipt, RlpEncodableReceipt, TxReceipt},
-    Typed2718,
-};
+use crate::receipt::{Eip658Value, RlpDecodableReceipt, RlpEncodableReceipt, TxReceipt};
 use alloc::{vec, vec::Vec};
 use alloy_eips::eip2718::Encodable2718;
 use alloy_primitives::{Bloom, Log};
-use alloy_rlp::{BufMut, Decodable, Encodable};
+use alloy_rlp::{BufMut, Decodable, Encodable, Header};
 use core::{borrow::Borrow, fmt};
 use derive_more::{DerefMut, From, IntoIterator};
+
+use super::Eip2718EncodableReceipt;
 
 /// Receipt containing result of transaction execution.
 #[derive(Clone, Debug, Default, PartialEq, Eq)]
@@ -72,7 +71,7 @@ where
     }
 }
 
-impl<T: Encodable> RlpEncodableReceipt for Receipt<T> {
+impl<T: Encodable> Receipt<T> {
     fn rlp_encoded_fields_length_with_bloom(&self, bloom: &Bloom) -> usize {
         self.status.length()
             + self.cumulative_gas_used.length()
@@ -86,9 +85,24 @@ impl<T: Encodable> RlpEncodableReceipt for Receipt<T> {
         bloom.encode(out);
         self.logs.encode(out);
     }
+
+    fn rlp_header_with_bloom(&self, bloom: &Bloom) -> Header {
+        Header { list: true, payload_length: self.rlp_encoded_fields_length_with_bloom(bloom) }
+    }
 }
 
-impl<T: Decodable> RlpDecodableReceipt for Receipt<T> {
+impl<T: Encodable> RlpEncodableReceipt for Receipt<T> {
+    fn rlp_encoded_length_with_bloom(&self, bloom: &Bloom) -> usize {
+        self.rlp_header_with_bloom(bloom).length_with_payload()
+    }
+
+    fn rlp_encode_with_bloom(&self, bloom: &Bloom, out: &mut dyn BufMut) {
+        self.rlp_header_with_bloom(bloom).encode(out);
+        self.rlp_encode_fields_with_bloom(bloom, out);
+    }
+}
+
+impl<T: Decodable> Receipt<T> {
     fn rlp_decode_fields_with_bloom(buf: &mut &[u8]) -> alloy_rlp::Result<ReceiptWithBloom<Self>> {
         let status = Decodable::decode(buf)?;
         let cumulative_gas_used = Decodable::decode(buf)?;
@@ -96,6 +110,25 @@ impl<T: Decodable> RlpDecodableReceipt for Receipt<T> {
         let logs = Decodable::decode(buf)?;
 
         Ok(ReceiptWithBloom { receipt: Self { status, cumulative_gas_used, logs }, logs_bloom })
+    }
+}
+
+impl<T: Decodable> RlpDecodableReceipt for Receipt<T> {
+    fn rlp_decode_with_bloom(buf: &mut &[u8]) -> alloy_rlp::Result<ReceiptWithBloom<Self>> {
+        let header = Header::decode(buf)?;
+        if !header.list {
+            return Err(alloy_rlp::Error::UnexpectedString);
+        }
+
+        let remaining = buf.len();
+
+        let this = Self::rlp_decode_fields_with_bloom(buf)?;
+
+        if buf.len() + header.payload_length != remaining {
+            return Err(alloy_rlp::Error::UnexpectedLength);
+        }
+
+        Ok(this)
     }
 }
 
@@ -234,21 +267,18 @@ impl<R: RlpDecodableReceipt> Decodable for ReceiptWithBloom<R> {
 
 impl<R> Encodable2718 for ReceiptWithBloom<R>
 where
-    R: RlpEncodableReceipt + Typed2718 + Send + Sync,
+    R: Eip2718EncodableReceipt + Send + Sync,
 {
     fn type_flag(&self) -> Option<u8> {
         (!self.receipt.is_legacy()).then_some(self.receipt.ty())
     }
 
     fn encode_2718(&self, out: &mut dyn BufMut) {
-        if let Some(ty) = self.type_flag() {
-            out.put_u8(ty);
-        }
-        self.encode(out);
+        self.receipt.eip2718_encode_with_bloom(&self.logs_bloom, out);
     }
 
     fn encode_2718_len(&self) -> usize {
-        self.type_flag().is_some() as usize + self.length()
+        self.receipt.eip2718_encoded_length_with_bloom(&self.logs_bloom)
     }
 }
